move const to common
