+++ internal/ui/overlay/manager_linux_common.go
@@ -5,6 +5,7 @@ package overlay
 import (
 	"image"
 	"os"
+	"strconv"
 	"strings"
 	"sync"
 	"unsafe"
@@ -20,6 +21,26 @@ import (
 	domainGrid "github.com/y3owk1n/neru/internal/core/domain/grid"
 	derrors "github.com/y3owk1n/neru/internal/core/errors"
 	"github.com/y3owk1n/neru/internal/core/ports"
+)
+
+const (
+	gridFillColor         uint32 = 0x18000000
+	gridMatchedFillColor  uint32 = 0x66465FBC
+	gridMatchedTextColor  uint32 = 0xFFF8FAFF
+	subgridBackground     uint32 = 0x40000000
+	subgridCellBackground uint32 = 0x10000000
+	badgePaddingX                = 10
+	badgeFontSize                = 14.0
+	badgeCharWidth               = 9
+	badgeHeight                  = 24
+	hexColorOpaque        uint32 = 0xFFFFFFFF
+	hexColorRepeatCount          = 2
+	subgridCols                  = 3
+	subgridRows                  = 3
+	subgridHalfPixel             = 0.5
+	subgridFontScale             = 0.7
+	subgridLineWidth             = 1
+	keyboardChanBuffer           = 64
 )
 
 type linuxOverlayBackend string
@@ -491,4 +512,26 @@ func resolveStickyIndicatorColors() overlayColors {
 		border:     parseHexColor(config.StickyModifiersBorderColorLight),
 		text:       parseHexColor(config.StickyModifiersTextColorLight),
 	}
+}
+
+func parseHexColor(value string) uint32 {
+	value = strings.TrimPrefix(strings.TrimSpace(value), "#")
+	switch len(value) {
+	case 3:
+		value = "FF" + strings.Repeat(string(value[0]), hexColorRepeatCount) +
+			strings.Repeat(string(value[1]), hexColorRepeatCount) +
+			strings.Repeat(string(value[2]), hexColorRepeatCount)
+	case 6:
+		value = "FF" + value
+	case 8:
+	default:
+		return hexColorOpaque
+	}
+
+	parsed, err := strconv.ParseUint(value, 16, 32)
+	if err != nil {
+		return hexColorOpaque
+	}
+
+	return uint32(parsed)
 }
